BACKPORT: Fix external command parsing bug.
The single argument StringList constructor only considers a tab to be a
valid separator. As such a space between a command and argument (e.g.
"convert.pl -x") is not treated as a separator, and the string is not
split properly. This results in execv calling a command that generally
doesn't exist (e.g. "convert.pl -x" rather than convert.pl).

Incidentally this change merges:
- Patch from ht://Dig ftp server: CommandParsing.0
- fedora: htdig-3.2.0-external_parsers part 1
- old htdig tree after 3.2.0-b6: first part of "patches from
  Jim Cole" (commit f0482c39bc87bd008783fc355f10270bec036210)

Link: https://github.com/roklein/htdig/commit/d0c8275fc6b251d94b721ba1a37cce75ee0ee706

Signed-off-by: Ilikara <3435193369@qq.com>